moved functions for creating matchups into own actions
