+++ src/Tournament/actions/find-identifier.js
@@ -1,0 +1,3 @@
+export const findIdentifier = (playerObj) => {
+  return playerObj?.guest_id ? playerObj?.guest_id : playerObj?.id
+}
