+++ src/Tournament/actions/player-arg-creator.js
@@ -1,0 +1,33 @@
+
+export function playerArgCreator(playerOppRef, refObj, scoreObject, actTourneyPlayers, playerBWTally) {
+  let identifier = null
+  let isActive = true
+  let playerArg = {}
+  //check if player is active and get identifier
+  if (isNaN(parseInt(playerOppRef))) {
+    identifier = playerOppRef
+    if (!actTourneyPlayers.find(ap => ap.guest_id === identifier)) {
+      isActive = false
+    }
+  }
+  else {
+    identifier = parseInt(playerOppRef)
+    if (!actTourneyPlayers.find(ap => ap.id === identifier)) {
+      isActive = false
+    }
+  }
+  if (isActive) {
+    playerArg = {
+      id: identifier,
+      score: scoreObject[identifier] || 0,
+      colors: playerBWTally,
+      avoid: refObj[identifier] ? refObj[identifier].filter(ref => ref !== 'bye') : []
+    }
+    if (refObj[identifier]) {
+      if (refObj[identifier].includes('bye') && playerArg.score > .5) {
+        playerArg.score--
+      }
+    }
+    return playerArg
+  }
+}
